requirements-dev.txt
# Development dependencies (extends production requirements)
-r requirements.txt

# Testing
pytest==9.0.2
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.5.0
pytest-timeout>=2.2.0
moto==5.1.21
hypothesis==6.151.9

# Code quality
black==26.1.0
flake8==7.3.0
mypy==1.19.1
isort>=5.13.0
pylint>=3.0.0
boto3-stubs[essential]>=1.34.0

# Security scanning
pip-audit>=2.7.0

# Pre-commit hooks
pre-commit>=3.6.0
